requirements.txt
dirtyjson>=1.0,<2
Pillow>=10.0,<11
reportlab>=3.6,<4
